+++ test/Acceptance/ExpectedClasses/Test6States.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Api;
+namespace Api\Schema;
 
 enum Test6States: string
 {